feat: move Navbar to components folder
